requirements.txt
Django~=5.0.3
djangorestframework~=3.14.0
httpx~=0.27.0
pydantic~=2.6.4
channels~=4.0.0